Clean up profile provider registrations
