+++ src/plugins/discover/public/context_awareness/profile_providers/register_profile_providers.ts
@@ -14,22 +14,15 @@ import type {
   RootProfileService,
 } from '../profiles';
 import type { BaseProfileProvider, BaseProfileService } from '../profile_service';
-import { createApacheErrorLogsDataSourceProfileProvider } from './apache_error_logs_data_source_profile';
-import { createAwsS3LogsDataSourceProfileProvider } from './aws_s3_logs_data_source_profile';
 import { exampleDataSourceProfileProvider } from './example_data_source_profile';
 import { exampleDocumentProfileProvider } from './example_document_profile';
 import { exampleRootProfileProvider } from './example_root_pofile';
-import { createK8ContainerLogsDataSourceProfileProvider } from './k8_container_logs_data_source_profile';
-import { createLogsDataSourceProfileProvider } from './logs_data_source_profile';
+import { createLogsDataSourceProfileProviders } from './logs_data_source_profile';
 import { createLogDocumentProfileProvider } from './log_document_profile';
-import { createNginxAccessLogsDataSourceProfileProvider } from './nginx_access_logs_data_source_profile';
-import { createNginxErrorLogsDataSourceProfileProvider } from './nginx_error_logs_data_source_profile';
 import {
   createProfileProviderServices,
   ProfileProviderServices,
 } from './profile_provider_services';
-import { createSystemLogsDataSourceProfileProvider } from './system_logs_data_source_profile';
-import { createWindowsLogsDataSourceProfileProvider } from './windows_logs_data_source_profile';
 
 export const registerProfileProviders = ({
   rootProfileService,
@@ -94,48 +87,13 @@ export const registerEnabledProfileProviders = <
 const extractProfileIds = (providers: Array<BaseProfileProvider<{}>>) =>
   providers.map(({ profileId }) => profileId);
 
-const createRootProfileProviders = (_providerServices: ProfileProviderServices) => {
-  return [] as RootProfileProvider[];
-};
+const createRootProfileProviders = (_providerServices: ProfileProviderServices) =>
+  [] as RootProfileProvider[];
 
-const createDataSourceProfileProviders = (providerServices: ProfileProviderServices) => {
-  const logsDataSourceProfileProvider = createLogsDataSourceProfileProvider(providerServices);
-  const systemLogsDataSourceProfileProvider = createSystemLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const k8ContainerLogsDataSourceProfileProvider = createK8ContainerLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const windowsLogsDataSourceProfileProvider = createWindowsLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const awsS3LogsDataSourceProfileProvider = createAwsS3LogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const nginxErrorLogsDataSourceProfileProvider = createNginxErrorLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const nginxAccessLogsDataSourceProfileProvider = createNginxAccessLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
-  const apacheErrorLogsDataSourceProfileProvider = createApacheErrorLogsDataSourceProfileProvider(
-    logsDataSourceProfileProvider
-  );
+const createDataSourceProfileProviders = (providerServices: ProfileProviderServices) => [
+  ...createLogsDataSourceProfileProviders(providerServices),
+];
 
-  return [
-    systemLogsDataSourceProfileProvider,
-    k8ContainerLogsDataSourceProfileProvider,
-    windowsLogsDataSourceProfileProvider,
-    awsS3LogsDataSourceProfileProvider,
-    nginxErrorLogsDataSourceProfileProvider,
-    nginxAccessLogsDataSourceProfileProvider,
-    apacheErrorLogsDataSourceProfileProvider,
-    logsDataSourceProfileProvider,
-  ];
-};
-
-const createDocumentProfileProviders = (providerServices: ProfileProviderServices) => {
-  const logsDocumentProfileProvider = createLogDocumentProfileProvider(providerServices);
-
-  return [logsDocumentProfileProvider];
-};
+const createDocumentProfileProviders = (providerServices: ProfileProviderServices) => [
+  createLogDocumentProfileProvider(providerServices),
+];
